Add Gml.WebApi.Tests project and initial unit test
This commit introduces the Gml.WebApi.Tests project, including the crucial .csproj file and global usings declaration. An initial unit test has been added within the file "Tests.cs" to assure that the get profile info functionality is correctly in place and this test will serve as the base for future testing implementations.