Reorganize Docker-related files in project
This update relocates the Dockerfile from the 'src' directory to the 'Gml.Web.Api' directory. In tandem with this change, the reference in the docker-compose file has been updated to reflect the new location. The .dockerignore file has also been removed from the Gml.Web.Api.csproj file following this restructuring.