Update to v0.1.0-beta4 (#88)
* Add Modrinth.Api and Modrinth.Api.Domains projects

Integrated the Modrinth.Api and Modrinth.Api.Domains projects into the solution file. Configurations for Debug and Release builds for these projects have also been added. This setup ensures they are correctly built along with the rest of the solution.

* Remove commented-out gml-kubek service from docker-compose.

The gml-kubek service configuration was commented out and is no longer needed. Removing it helps clean up the docker-compose file and ensures it only contains active and necessary configurations.