Update to v0.1.0-rc1 (#89)
* Add Modrinth.Api and Modrinth.Api.Domains projects

Integrated the Modrinth.Api and Modrinth.Api.Domains projects into the solution file. Configurations for Debug and Release builds for these projects have also been added. This setup ensures they are correctly built along with the rest of the solution.

* Remove commented-out gml-kubek service from docker-compose.

The gml-kubek service configuration was commented out and is no longer needed. Removing it helps clean up the docker-compose file and ensures it only contains active and necessary configurations.

* Update volume paths and remove Filebrowser service

Change the volume paths for GmlBackend, Minio, and TextureService to specific directories under ./data to enhance clarity and organization. Removed the Filebrowser service as it is no longer required in the configuration.

* Update volume mappings and remove unused filebrowser service

Updated volume mappings to use specific directories instead of the shared volume for the GmlBackend, Minio, and TextureService containers. Removed the unused filebrowser service and its configuration to clean up the docker-compose file for production environments.

* Set UID and GID for MinIO service in Docker Compose files.

Added user settings to `docker-compose-prod.yml` and `docker-compose.yml` to specify `UID` and `GID` for the MinIO service. This ensures correct file permissions and enhances security by running MinIO under non-root user IDs.

* Update Update submodule link Gml.Web.Api

* Update submodule link Gml.Web.Client

* Update submodule link Gml.Web.Skin.Service