Update to v0.1.0-rc2-hotfix1 (#92)
* Update submodule link Gml.Web.Client

* Update submodule link Gml.Web.Skin.Service

* Update submodule link Gml.Web.Api

* Update submodule link Gml.Core

* Update Update submodule link Gml.Web.Api

* Update submodule link Gml.Web.Api

* Add SERVICE_TEXTURE_ENDPOINT to environment configuration

Included SERVICE_TEXTURE_ENDPOINT in docker-compose.yml and docker-compose-prod.yml files to pass the variable to the respective service. Updated the .env.example file to provide a default value for SERVICE_TEXTURE_ENDPOINT.

* Update submodule link Gml.Web.Api

* Update submodule link Gml.Web.Client

* Add 'data' directory to .gitignore

This change updates the .gitignore file to exclude the 'data' directory. This ensures that any files or folders named 'data' will not be tracked by Git, helping to keep the repository clean and focused on source code and relevant resources.

* Update submodule link Gml.Web.Api

* Update submodule link Gml.Web.Api

* Update submodule link Gml.Web.Api

* Update submodule link Gml.Web.Client

* Update submodule link Gml.Web.Skin.Service